requirements.txt
asgiref==3.7.2
Django==4.2.6
django-cors-headers==4.2.0
django-environ==0.11.2
djangorestframework==3.14.0
pytz==2023.3.post1
PyYAML==6.0.1
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
